pyproject.toml
[tool.poetry]
name = "notes-api"
version = "0.1.0"
description = ""
authors = ["Chloe Nott <chloe@nott.email>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
